getting sensor right

+++ plax_platform.py
@@ -686,7 +686,7 @@ class Parallax_Platform(Hardware):
             # 1/(steps/s) = (inter-step interval in s)*1000000 (to micros)
             # delay_dur = (inter-step interval) - (pulse_dur)
             # TODO: clip delay dur for asymptotically slow velocities
-            delay_dur = np.round(((1 / (abs(velocity) / self.MM_PER_STEP) * 1000000) - self.pulse_dur))
+            delay_dur = np.round(((1 / (abs(velocity) / self.MM_PER_STEP) * 1000000) - self.pulse_dur)).astype(int)
             if delay_dur < 0:
                 self.logger.warning(
                     f"Could not set velocity to {velocity}, pulse dur {self.pulse_dur} is too long to go this fast!")
